Refactor PagedList components to be independent from the lists it renders
