+++ lang/main.json
@@ -55,6 +55,7 @@
         "go": "GO",
         "join": "JOIN",
         "privacy": "Privacy",
+        "recentList": "History",
         "roomname": "Enter room name",
         "roomnameHint": "Enter the name or URL of the room you want to join. You may make a name up, just let the people you are meeting know it so that they enter the same name.",
         "sendFeedback": "Send feedback",
