+++ react/features/base/react/components/native/index.js
@@ -3,6 +3,7 @@ export { default as Header } from './Header';
 export { default as NavigateSectionList } from './NavigateSectionList';
 export { default as Link } from './Link';
 export { default as LoadingIndicator } from './LoadingIndicator';
+export { default as PagedList } from './PagedList';
 export { default as Pressable } from './Pressable';
 export { default as SideBar } from './SideBar';
 export { default as Text } from './Text';
